internal/server: gate UI embed behind embed_ui build tag
Replace the ui_dist placeholder.txt hack with a build-tag guard so plain
`go build`/`go test` no longer require a populated ui_dist directory. The
real embed only happens when compiled with `-tags embed_ui`.

- embed.go (//go:build embed_ui) holds the //go:embed all:ui_dist directive
- embed_notag.go (//go:build !embed_ui) serves an empty FS for tests
- Makefile build targets pass -tags embed_ui and drop the placeholder touch
- .goreleaser.yaml builds add the embed_ui tag so releases ship the UI

Docker images consume the goreleaser binary, and CI runs tests without the
tag, so no changes are needed there.
