+++ .gitignore
@@ -6,5 +6,5 @@ dist/
 .DS_Store
 .dev/
 
-# UI build output; placeholder.txt is kept so the go:embed succeeds.
+# UI build output; embedded via the `embed_ui` build tag (internal/server/embed.go)
 internal/server/ui_dist/*
